[CHANGE] move Billboard Helper to main helper path
